pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo2readme"
version = "0.1.0"
authors = [
  { name = "Tiago Silva" },
]
description = "A utility for managing and organizing notes"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "tqdm",
    "openai",
    "python-dotenv"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
repo2readme = "repo2readme.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/repo2readme"]

[tool.hatch.build]
include = [
  "src/repo2readme/.env.example",
]